+++ src/lib/st-pagination/st-pagination.component.spec.ts
@@ -10,10 +10,10 @@
  */
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { StPaginationComponent } from './st-pagination.component';
-import { NO_ERRORS_SCHEMA, SimpleChange } from '@angular/core';
+import { NO_ERRORS_SCHEMA, SimpleChange, DebugElement } from '@angular/core';
+import { By } from '@angular/platform-browser';
 
 describe('StPaginationComponent', () => {
-
    let component: StPaginationComponent;
    let fixture: ComponentFixture<StPaginationComponent>;
 
@@ -21,7 +21,6 @@ describe('StPaginationComponent', () => {
       TestBed.configureTestingModule({
          declarations: [StPaginationComponent],
          schemas: [NO_ERRORS_SCHEMA]
-
       })
          .compileComponents();  // compile template and css
    }));
@@ -29,54 +28,106 @@ describe('StPaginationComponent', () => {
    beforeEach(() => {
       fixture = TestBed.createComponent(StPaginationComponent);
       component = fixture.componentInstance;
+   });
 
+   it('Should be initialized correctly', () => {
+      (fixture.elementRef.nativeElement as HTMLElement).id = null;
+      fixture.detectChanges();
+
+      expect(component.currentPage).toEqual(1);
+      expect(component.perPage).toEqual(20);
+      expect(component.total).toBeUndefined();
+      expect(component.label).toEqual({
+         element: 'Rows', perPage: 'per page', of: 'of'
+      });
+      expect(component.perPageOptions).toEqual([
+         { value: 20, showFrom: 0 }, { value: 50, showFrom: 50 }, { value: 200, showFrom: 200 }
+      ]);
+
+      expect(component.disableNextButton).toBeFalsy();
+      expect(component.disablePrevButton).toBeTruthy();
+      expect(component.firstItem).toEqual(1);
+      expect(component.lastItem).toEqual(20);
+      expect(component.items).toEqual([]);
+      expect(component.selectedItem).toBeUndefined();
+
+      expect(component.hasOptions).toBeFalsy();
+      expect(component.paginationId).toBeNull();
+      expect(component.selectId).toBeNull();
+      expect(component.buttonPrevId).toBeNull();
+      expect(component.buttonNextId).toBeNull();
+   });
+
+   it('Should propagate id to internal elements', () => {
+      const id: string = 'test-id';
+      (fixture.elementRef.nativeElement as HTMLElement).id = id;
+      component.perPage = 20;
+      component.total = 500;
+      fixture.detectChanges();
+
+      const select: HTMLElement = fixture.debugElement.query(By.css('st-select')).nativeElement;
+      const buttons: DebugElement[] = fixture.debugElement.queryAll(By.css('.button-toolbar'));
+
+      expect(component.paginationId).toEqual(id);
+
+      expect(component.selectId).toEqual(`${id}-select`);
+      expect(select.getAttribute('id')).toEqual(`${id}-select`);
+
+      expect(component.buttonPrevId).toEqual(`${id}-prev`);
+      expect(buttons[0].nativeElement.getAttribute('id')).toContain(`${id}-prev`);
+
+      expect(component.buttonNextId).toEqual(`${id}-next`);
+      expect(buttons[1].nativeElement.getAttribute('id')).toContain(`${id}-next`);
    });
 
    describe('When insert input perPage', () => {
 
       it('should not show dropdown', () => {
          component.perPage = 20;
-         component.total = 50;
-         expect(component.showItemsPerPage()).toBeFalsy();
+         component.total = 20;
+         fixture.detectChanges();
+
+         const select: DebugElement = fixture.debugElement.query(By.css('st-select'));
+
+         expect(select).toBeNull();
       });
 
       it('should show dropdown', () => {
          component.perPage = 20;
          component.total = 1000;
-         expect(component.showItemsPerPage()).toBeTruthy();
-      });
-
-
-      it('should not show the dropdown menu', () => {
-         component.perPage = 20;
-         component.total = 40;
          fixture.detectChanges();
-         expect(component.showItemsPerPage()).toBeFalsy();
-      });
 
+         const select: DebugElement = fixture.debugElement.query(By.css('st-select'));
+
+         expect(select).toBeDefined();
+      });
    });
 
    describe('When insert input perPageOptions', () => {
 
-      it('should be items equal to options insert by user', () => {
-
-         component.perPageOptions = [10, 20, 30];
-         fixture.detectChanges();
-         expect(component.items[0].value).toBe(10);
-      });
-
       it('should be items equal to default per page options', () => {
+         component.perPage = 20;
+         component.total = 1000;
          fixture.detectChanges();
-         expect(component.items[0].value).toBe(20);
+
+         expect(component.items.length).toEqual(3);
       });
 
+      it('should be items equal to options insert by user', () => {
+         component.perPage = 20;
+         component.total = 1000;
+         component.perPageOptions = [
+            { value: 20 }, { value: 50 }, { value: 100 }, { value: 150 }, { value: 200 }
+         ];
+         fixture.detectChanges();
 
+         expect(component.items.length).toEqual(5);
+      });
    });
 
    describe('When update the pagination', () => {
 
       it('should be the next page', () => {
-
          component.perPage = 20;
          component.total = 100;
          component.currentPage = 2;
@@ -88,7 +139,6 @@ describe('StPaginationComponent', () => {
       });
 
       it('should be the prev page', () => {
-
          component.perPage = 20;
          component.total = 100;
          component.currentPage = 2;
@@ -100,7 +150,6 @@ describe('StPaginationComponent', () => {
       });
 
       it('should be disable the next button', () => {
-
          component.perPage = 20;
          component.total = 40;
          component.currentPage = 1;
@@ -112,7 +161,6 @@ describe('StPaginationComponent', () => {
       });
 
       it('should be disable the prev button', () => {
-
          component.perPage = 20;
          component.total = 40;
          component.currentPage = 2;
@@ -128,7 +176,6 @@ describe('StPaginationComponent', () => {
    describe('when component is update', () => {
 
       it('should generate new item for dropdown', () => {
-
          component.perPage = 20;
          component.total = 50;
          fixture.detectChanges();
@@ -139,12 +186,47 @@ describe('StPaginationComponent', () => {
          fixture.detectChanges();
 
          expect(component.items.length).toBe(3);
-
-
       });
 
+      it('should change page attributes for page change', () => {
+         component.perPage = 20;
+         component.total = 45;
+         fixture.detectChanges();
 
+         fixture.componentInstance.currentPage = 3;
+
+         component.ngOnChanges({ currentPage: new SimpleChange(1, 3, false) });
+         fixture.detectChanges();
+
+         expect(component.disableNextButton).toBeTruthy();
+         expect(component.disablePrevButton).toBeFalsy();
+         expect(component.firstItem).toEqual(41);
+         expect(component.lastItem).toEqual(45);
+      });
    });
 
+   describe('when component select a new perpage', () => {
 
+      it('should update to first page', () => {
+         component.perPage = 20;
+         component.total = 500;
+         component.currentPage = 3;
+         fixture.detectChanges();
+
+         expect(component.disableNextButton).toBeFalsy();
+         expect(component.disablePrevButton).toBeFalsy();
+         expect(component.firstItem).toEqual(41);
+         expect(component.lastItem).toEqual(60);
+         expect(component.selectedItem.value).toEqual(20);
+
+         fixture.componentInstance.onChangePerPage(50);
+         fixture.detectChanges();
+
+         expect(component.disableNextButton).toBeFalsy();
+         expect(component.disablePrevButton).toBeTruthy();
+         expect(component.firstItem).toEqual(1);
+         expect(component.lastItem).toEqual(50);
+         expect(component.selectedItem.value).toEqual(50);
+      });
+   });
 });