+++ src/lib/st-pagination/st-pagination.pipe.ts
@@ -12,7 +12,6 @@ import { Pipe, PipeTransform } from '@angular/core';
 
 import { StPaginationService } from './st-pagination.service';
 
-
 @Pipe({
    name: 'paginate',
    pure: false
@@ -25,6 +24,4 @@ export class StPaginationPipe implements PipeTransform {
    transform(items: any[], args: any): any {
       return this.service.newPage(items, args.currentPage, args.perPage);
    }
-
-
 }